Fix signedness issues in C client
- http.c: Use explicit size_t casts to avoid signed/unsigned mixing
- speedtest.c: Cast payload_len to int64_t for bytes_to_mbps()
- Use (size_t)n casts when adding ssize_t to size_t counters

+++ netspeed.c/src/speedtest.c
@@ -221,12 +221,12 @@ double measure_upload(http_conn_t *conn, const char *base_url,
         return -1;
     }
 
-    double mbps = bytes_to_mbps(payload_len, duration_ms);
+    double mbps = bytes_to_mbps((int64_t)payload_len, duration_ms);
 
     if (sample) {
         sample->ts = timing_now_ms();
         sample->direction = "upload";
-        sample->size_bytes = payload_len;
+        sample->size_bytes = (int64_t)payload_len;
         sample->duration_ms = duration_ms;
         sample->mbps = mbps;
         sample->profile = profile;
